Features data load fix (#108)
Summary:
Depends on https://github.com/facebookresearch/Ego4d/issues/92

With no seek (as in the code), the number of requested frames retrieved will be less than expected. Specifically the first stride frames will be not included in every window fed into the model. This is some serious negligence on my part (I really don't know what I was thinking at the time...). SlowFast features should be unaffected. This went un-noticed due to the transform: `UniformTemporalSubsample` which padded the frames to the window size - which was previously there due to EncodedVideo being imprecise previously.

This solution is a bit hacky, and does not perform a seek. Seeking simply doesn't work on the FAIR cluster with PyAV (have yet to try locally).

Instead I:
1. Iterate over each frame on each worker, caching them into a fixed window buffer
2. On the next iteration: iterate over the cache window, checking if there's any frames overlapping with the current window to retrieve
3. Iterate over the container for more possible frames

Note this does depend on the fact that the dataloader will request examples in order (which it does).

Alternative options are not feasible at this point.. I did also try decord (via compilation) and I cannot get it to work with the FAIR cluster on our set of videos.

This doesn't support audio, so I'll need to fix my other PRs.

Pull Request resolved: https://github.com/facebookresearch/Ego4d/pull/108

Reviewed By: ebyrne

Differential Revision: D36417513

Pulled By: miguelmartin75

fbshipit-source-id: 264edf8cba15ccf2a3abd2249334ffcae85ff5ec