Fix circular import error due to types.py (#231)
Summary:
This pull request addresses a circular import issue caused by a conflict with the built-in Python types module. Currently, the `types.py `file in our project is conflicting with the standard types module, leading to an `AttributeError` when using re.compile. To resolve this, we propose renaming `types.py` to `manifest.py.` Additionally, we have made necessary updates to `validate.py` and `ffmpeg_utils.py` to accommodate these changes.

Pull Request resolved: https://github.com/facebookresearch/Ego4d/pull/231

Reviewed By: devanshk

Differential Revision: D48629219

Pulled By: miguelmartin75

fbshipit-source-id: 8ba01ce34996b98890f62aaa7121ebbf8394c6e5